refactor: domain 내부패키지 수정
도메인패키지 안의 menu 패키지 제거하였다. 굳이 menu 패키지를 따로 둘 필요가 없다고 생각하여 menu 패키지 안의 enum은 domain으로 이동하였다.
