+++ src/main/java/subway/controller/SearchingPathController.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 import subway.common.ErrorCustomException;
 import subway.common.ErrorMessage;
 import subway.domain.PathRepository;
-import subway.domain.menu.Menu;
+import subway.domain.Menu;
 import subway.view.menu.MenuInputManager;
-import subway.domain.menu.MenuKeys;
+import subway.domain.MenuKeys;
 import subway.view.menu.MenuOutputManager;
 import subway.view.path.PathInputManager;
 import subway.view.path.PathOutputManager;
