+++ src/main/java/subway/domain/Menu.java
@@ -1,4 +1,4 @@
-package subway.domain.menu;
+package subway.domain;
 
 import java.util.ArrayList;
 import java.util.Arrays;
