+++ src/main/java/subway/view/menu/MenuInputManager.java
@@ -3,7 +3,7 @@ package subway.view.menu;
 import java.util.Scanner;
 import subway.common.ErrorCustomException;
 import subway.common.ErrorMessage;
-import subway.domain.menu.Menu;
+import subway.domain.Menu;
 
 public class MenuInputManager {
     private static final String INPUT_NEEDED = " 중에서 입력해 주세요.";
