+++ src/main/java/subway/view/menu/MenuOutputManager.java
@@ -1,7 +1,7 @@
 package subway.view.menu;
 
 import subway.common.GuideMessage;
-import subway.domain.menu.Menu;
+import subway.domain.Menu;
 
 public class MenuOutputManager {
     public static final String WHICH_MENU = "원하는 기능을 선택하세요.";
